Refactor compose-app package.
