+++ compose-app/src/main/java/com/mapbox/maps/compose/testapp/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.compose.ui.theme
+package com.mapbox.maps.compose.testapp.ui.theme
 
 import androidx.compose.ui.graphics.Color
