+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/internal/MapPreviewPlaceholder.kt
@@ -2,12 +2,14 @@ package com.mapbox.maps.extension.compose.internal
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.unit.dp
+import com.mapbox.maps.MapInitOptions
 
 /**
  * A place holder to show the Map in Android Studio Preview.
@@ -16,6 +18,7 @@ import androidx.compose.ui.unit.dp
 @JvmSynthetic
 internal fun MapPreviewPlaceHolder(
   modifier: Modifier,
+  mapInitOptions: MapInitOptions,
   placeholderColor: Color = Color(0xFFD2D2D9),
   strokeWidthInDp: Int = 5,
   strokeColor: Color = Color.Blue
@@ -41,4 +44,6 @@ internal fun MapPreviewPlaceHolder(
       color = strokeColor
     )
   }
+
+  Text(text = mapInitOptions.toString())
 }
